ubench: add initial micro-benchmark framework
currently, the micro-benchmarks do not seem to be working; when
inspecting the SASS obtained from cuobjdump, the benchmark kernel
contains to FMA or any other instructions related to the ring operation
being benchmark. This needs investigation, but this commits provides a
framework for what the benchmark architecture should look like.